Update markdown style
I've made no changes to the content, only to the style in which the markdown was written. This should make editing the source further in the future a tad more ergonomic. The forward facing changes for the reader at this point should be minimal.

The primary change in style is the move to Asciidoctor's "One sentence per line" approach, aka "ventilated prose". The benefits of this style are explained further here:
https://asciidoctor.org/docs/asciidoc-recommended-practices/#one-sentence-per-line

In addition to that, I've included more spacing between elements, and replaced the use of hyphens with the appropriate em-dash HTML code.

Next, I'll create a second PR to improve the grammar and correct typos.